Validate port and proxied_path
When subclassing LocalProxy, it's easy for other packages
to accidentally pass non-integers as ports or proxied_paths that
don't have a / at start. We used to try to correct for this,
but it's better to error. The subclasses might need to
adjust their routes to include the / in their route regex.